fix(review): close advisor gate review gaps
